+++ dataworkspace/dataworkspace/apps/core/views.py
@@ -61,14 +61,7 @@ class SupportView(FormView):
     def form_valid(self, form):
         cleaned = form.cleaned_data
         ticket_id = create_support_request(
-            self.request.user,
-            cleaned['email'],
-            cleaned['message'],
-            attachments=[
-                cleaned[x]
-                for x in ['attachment1', 'attachment2', 'attachment3']
-                if cleaned[x] is not None
-            ],
+            self.request.user, cleaned['email'], cleaned['message']
         )
         return HttpResponseRedirect(
             reverse('support-success', kwargs={'ticket_id': ticket_id})